drmgr: ensure PCI cleanup is done before offlining dev
As part of 4f542616, functionality was added to support add/remove
PCI devices without relying on rpaphp hotplug module. The sysfs
write used in the remove path initiates an asynchronous removal of
PCI device from kernel PCI subsystem, and eventually from the sysfs
tree. Afterward, we set the corresponding DRC for the device to
ISOLATED state so host can complete device removal.

To avoid the situation where the asychronous removal has not yet
completed by the time drmgr sets the state to ISOLATED, and simple
sleep of 3 seconds was added. This seems to have worked out okay
for virtio devices, but in the case of other devices, like PCI
passthrough of a Mellanox CX4 virtual function, the removal can take
up to 20 seconds.

When this happens, we hit a situation where the host reclaims the
devices while the guest is still attempting to access it in it's
cleanup path, which leads to guest crashes.

Fix this by polling the PCI devices sysfs path and not completing
the removal until the sysfs path is removed as part of device cleanup.

Signed-off-by: Michael Roth <mdroth@linux.vnet.ibm.com>