drmgr: avoid loading rpadlpar_io/rpaphp when -v flags is used
The -v / opts->pci_virtio flag was added to provide an alternative way
to scan/remove PCI devices which doesn't rely on rpaphp. That was a
workaround for a bug induced by QEMU emulating PHBs with multiple
hotpluggable slots, which fails to register slots beyond the first due
to rpaphp no longer accounting for the possibility of there being more
than 1 slot per PHB. This leads to warnings of the form:

  kernel: rpaphp: pci_hp_register failed with error -16
  kernel: rpaphp: pci_hp_register failed with error -16
  ...

For most guest distros this has resulted in the module not being
loaded by default to avoid the errors showing up.

However, as of a9b4e4e, we now load automatically load the module as
needed, but we don't take opts->pci_virtio into account when
determining if it's needed, leading to PCI hotplug operations to
KVM guests now resulting into rpaphp being loaded and
generating the errors above when drmgr is invoked.

Fix this by avoiding modprobe of rpadlpar_io/rpaphp when
opts->pci_virtio is set.

Signed-off-by: Michael Roth <mdroth@linux.vnet.ibm.com>